requirements.txt
feedparser==6.0.10
discord.py==2.3.2
tweepy==4.14.0
aiohttp==3.8.6
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4>=4.12.0
aiosqlite>=0.19.0
atproto>=0.0.40

